requirements.txt
anthropic>=0.69.0
google-genai>=1.0.0
openai>=2.0.0
pandas>=2.2.0
plotly>=5.24.0
pydantic>=2.8.0
python-dotenv>=1.0.1
streamlit>=1.39.0
pytest>=8.3.0
